Issue 158: Moved SoftLayerParserModule to right package
