+++ sandbox-providers/softlayer/src/test/java/org/jclouds/softlayer/parse/ParseBadVirtualGuest.java
@@ -23,9 +23,7 @@ import com.google.inject.Injector;
 import org.jclouds.date.internal.SimpleDateFormatDateService;
 import org.jclouds.json.BaseItemParserTest;
 import org.jclouds.json.config.GsonModule;
-import org.jclouds.softlayer.compute.config.SoftLayerParserModule;
-import org.jclouds.softlayer.domain.Datacenter;
-import org.jclouds.softlayer.domain.OperatingSystem;
+import org.jclouds.softlayer.config.SoftLayerParserModule;
 import org.jclouds.softlayer.domain.PowerState;
 import org.jclouds.softlayer.domain.VirtualGuest;
 import org.testng.annotations.Test;
